Add method to save list of package versions to a csv file
Changelog
=========

A new function, export_package_versions(), was defined in the
config_rstudio_server.R source file. This function, when run, builds a table
of packages and currently installed versions and saves it to a csv file named
packages.csv. This file can be used to validate that all the necessary packages
have been installed for CDS 101 and CDS 102.